New version of bigop_tactics.v.
Summary:
- Add tactic support for eq_bigl (for rewriting bigop predicates).
- Add a shortcut: big := bigop _ _ _.
- Rephrase the tactic notations.
  Before:
    underbig (bigop _ _ _) i Hi rewrite lem.
  Now:
    under big i Hi rewrite lem.
- Improve the Ltac tactics so the bigop variable ("i" or so) is kept,
  if specified.

Note: this file correspond to the first commit of repo
https://github.com/erikmd/ssr-under-tac/commit/e2aeee0

Since then it has been significantly improved thanks to Cyril Cohen's
advice, but the latest version of my tactic only work with the
development version of MathComp, which contains a few bugs related to
the ssrpattern tactic. As soon as both issues
https://github.com/math-comp/math-comp/issues/61 and
https://github.com/math-comp/math-comp/issues/62 are resolved, I will
submit a PR for integrating https://github.com/erikmd/ssr-under-tac in
MathComp (and upgrade validsdp accordingly).