Improve Ltac error reporting.
Use a lazymatch-goal instead of match-goal for implementing do_sdp, so
soswitness' possible failures will be properly propagated to the user.

Regarding Ltac error reporting, the only "drawback" of current code is
in the case when ValidSDP would fail to verify the generated witnesses
(as the failure message would be very verbose in this case).