Add tactic check_unexpected_case
It triggers a failure in case get_comp_poly.aux2 is called with a wrong f0
(such as Rplus) and this happens to occur when an intermediate polynomial
has been defined using Let, not Definition.
This issue is due to function list_idx which doesn't succeed in comparing
the generated evars, cf. the message:
?dummy0 doesn't_belong_to [:: ?dummy; ?dummy0]